requirements.txt
flask
transformers
torch
pillow
torchvision